requirements.txt
absl-py==0.9.0
asgiref==3.2.5
astor==0.8.1
cachetools==4.0.0
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
defusedxml==0.6.0
Django==2.1
django-allauth==0.39.1
django-ckeditor==5.6.1
django-js-asset==1.2.2
django-model-utils==3.1.2
django-mptt==0.10.0
django-notifications-hq==1.5.0
django-password-reset==2.0
django-taggit==0.23.0
gast==0.2.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.8
jsonfield==2.0.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==2.6.11
mysqlclient==1.4.6
numpy==1.18.1
oauthlib==3.0.1
opencv-contrib-python==4.2.0.32
opt-einsum==3.2.0
Pillow==5.3.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.2.0
PyJWT==1.7.1
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2018.5
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scipy==1.4.1
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
system==0.1.16
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==1.0.0
wrapt==1.12.1
